requirements.txt
alembic==1.13.1
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.29.0
bcrypt==4.0.1
certifi==2025.7.9
cffi==1.17.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
et_xmlfile==2.0.0
Faker==37.4.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
litestar==2.16.0
litestar-htmx==0.5.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgspec==0.19.0
multidict==6.6.3
multipart==1.2.1
openpyxl==3.1.5
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
polyfactory==2.22.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.19.2
pytest==8.2.1
pytest-asyncio==0.23.6
python-dotenv==1.1.1
python-jose==3.3.0
PyYAML==6.0.2
rich==14.0.0
rich-click==1.8.9
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.47.2
typing_extensions==4.14.1
tzdata==2025.2
uvicorn==0.35.0
watchfiles==1.1.0
websockets==15.0.1
